+++ src/handlers/championships/sockets.rs
@@ -8,7 +8,7 @@ use hyper::StatusCode;
 use std::sync::Arc;
 
 #[inline(always)]
-pub async fn active_sockets(State(state): State<UserState>) -> AppResult<Json<Vec<i64>>> {
+pub async fn active_sockets(State(state): State<UserState>) -> AppResult<Json<Vec<i32>>> {
     let sockets = state.f123_service.active_sockets().await;
     Ok(Json(sockets))
 }
@@ -16,7 +16,7 @@ pub async fn active_sockets(State(state): State<UserState>) -> AppResult<Json<Ve
 #[inline(always)]
 pub async fn start_socket(
     State(state): State<UserState>,
-    Path(championship_id): Path<i64>,
+    Path(championship_id): Path<i32>,
 ) -> AppResult<Response> {
     let championship = state.championship_repository.find(&championship_id).await?;
 
@@ -31,7 +31,7 @@ pub async fn start_socket(
 #[inline(always)]
 pub async fn stop_socket(
     State(state): State<UserState>,
-    Path(championship_id): Path<i64>,
+    Path(championship_id): Path<i32>,
 ) -> AppResult<Response> {
     state.f123_service.stop_socket(championship_id).await?;
